Move reexec logic out of lib/bpf/common_test.go
