+++ lib/bpf/common_test.go
@@ -17,10 +17,6 @@ limitations under the License.
 package bpf
 
 import (
-	"io"
-	"net/http"
-	"os"
-	osexec "os/exec"
 	"testing"
 	"time"
 
@@ -29,49 +25,7 @@ import (
 
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/service/servicecfg"
-	"github.com/gravitational/teleport/lib/utils"
 )
-
-const (
-	// reexecInCGroupCmd is a cmd used to re-exec the test binary and call arbitrary program.
-	reexecInCGroupCmd = "reexecCgroup"
-	// networkInCgroupCmd is a cmd used to re-exec the test binary and make HTTP call.
-	networkInCgroupCmd = "networkCgroup"
-)
-
-func TestMain(m *testing.M) {
-	utils.InitLoggerForTests()
-
-	// Check if the re-exec was requested.
-	if len(os.Args) == 3 {
-		var err error
-
-		switch os.Args[1] {
-		case reexecInCGroupCmd:
-			// Get the command to run passed as the 3rd argument.
-			cmd := os.Args[2]
-
-			err = waitAndRun(cmd)
-		case networkInCgroupCmd:
-			// Get the endpoint to call.
-			endpoint := os.Args[2]
-
-			err = callEndpoint(endpoint)
-		default:
-			os.Exit(2)
-		}
-
-		if err != nil {
-			// Something went wrong, exit with error.
-			os.Exit(1)
-		}
-
-		// The rexec was handled and nothing bad happened.
-		os.Exit(0)
-	}
-
-	os.Exit(m.Run())
-}
 
 // TestCheckAndSetDefaults makes sure defaults are set when the user does not
 // provide values for the page sizes and hard coded values (like zero or a
@@ -132,44 +86,4 @@ func TestBPFConfig_CheckAndSetDefaults(t *testing.T) {
 			}
 		})
 	}
-}
-
-// waitAndRun wait for continue signal to be generated an executes the
-// passed command and waits until returns.
-func waitAndRun(cmd string) error {
-	if err := waitForContinue(); err != nil {
-		return err
-	}
-
-	return osexec.Command(cmd).Run()
-}
-
-// callEndpoint wait for continue signal to be generated an executes HTTP GET
-// on provided endpoint.
-func callEndpoint(endpoint string) error {
-	if err := waitForContinue(); err != nil {
-		return err
-	}
-
-	resp, err := http.Get(endpoint)
-	if resp != nil {
-		// Close the body to make our linter happy.
-		_ = resp.Body.Close()
-	}
-
-	return err
-}
-
-// waitForContinue opens FD 3 and waits the signal from parent process that
-// the cgroup is being observed and the even can be generated.
-func waitForContinue() error {
-	waitFD := os.NewFile(3, "/proc/self/fd/3")
-	defer waitFD.Close()
-
-	buff := make([]byte, 1)
-	if _, err := waitFD.Read(buff); err != nil && err != io.EOF {
-		return err
-	}
-
-	return nil
 }
